move note editor
