+++ entrypoints/popup/components/notes/__tests__/useNoteEditor.test.tsx
