+++ entrypoints/popup/components/notes/useNoteEditor.ts
